requirements.txt
fastapi==0.109.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
openai-whisper==20231117
torch==2.8.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiofiles==23.2.1
